requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
python-telegram-bot[job-queue]>=20.6
python-dotenv>=1.1.1
